+++ tests/pyincore/analyses/joplinempiricalbuildingrestoration/test_joplinempiricalbuildingrestoration.py
@@ -2,7 +2,7 @@
 
 from pyincore import IncoreClient, FragilityService, MappingSet
 from pyincore.analyses.buildingdamage import BuildingDamage
-from pyincore.analyses.joplinempiricalrestoration import JoplinEmpiricalRestoration
+from pyincore.analyses.joplinempiricalbuildingrestoration import JoplinEmpiricalBuildingRestoration
 import pyincore.globals as pyglobals
 
 
@@ -39,9 +39,9 @@ def run_with_base_class():
     building_dmg_result = bldg_dmg.get_output_dataset("ds_result")
     building_dmg_result.get_dataframe_from_csv()
 
-    restoration = JoplinEmpiricalRestoration(client)
+    restoration = JoplinEmpiricalBuildingRestoration(client)
     # Building final target level dataset defining custom final restoration values
-    building_fl_id = "61a68979d5b02930aa43ae39"
+    building_fl_id = "6629304a5da5fd22b250a66b"
 
     restoration.load_remote_input_dataset("buildings", bldg_dataset_id)
     # restoration.load_remote_input_dataset("building_dmg", building_dmg_result)